requirements.txt
fastapi
uvicorn
python-dotenv
openai
supabase
langchain
langchain-openai
langchain-community
tiktoken
requests
